+++ src/icalendar/cal/event.py
@@ -9,6 +9,7 @@ from typing import TYPE_CHECKING, Optional, Sequence
 from icalendar.attr import (
     X_MOZ_LASTACK_property,
     X_MOZ_SNOOZE_TIME_property,
+    attendees_property,
     categories_property,
     class_property,
     color_property,
@@ -26,6 +27,7 @@ from icalendar.attr import (
     rdates_property,
     rrules_property,
     sequence_property,
+    status_property,
     summary_property,
     transparency_property,
     uid_property,
@@ -38,7 +40,7 @@ from icalendar.tools import is_date
 
 if TYPE_CHECKING:
     from icalendar.alarms import Alarms
-    from icalendar.enums import CLASS, TRANSP
+    from icalendar.enums import CLASS, STATUS, TRANSP
     from icalendar.prop import vCalAddress
 
 
@@ -384,11 +386,14 @@ class Event(Component):
     priority = priority_property
     contacts = contacts_property
     transparency = transparency_property
+    status = status_property
+    attendees = attendees_property
 
     @classmethod
     def new(
         cls,
         /,
+        attendees: Optional[list[vCalAddress]] = None,
         categories: Sequence[str] = (),
         classification: Optional[CLASS] = None,
         color: Optional[str] = None,
@@ -404,6 +409,7 @@ class Event(Component):
         sequence: Optional[int] = None,
         stamp: Optional[date] = None,
         start: Optional[date | datetime] = None,
+        status: Optional[STATUS] = None,
         transparency: Optional[TRANSP] = None,
         summary: Optional[str] = None,
         uid: Optional[str | uuid.UUID] = None,
@@ -414,6 +420,7 @@ class Event(Component):
         This creates a new Event in accordance with :rfc:`5545`.
 
         Arguments:
+            attendees: The :attr:`attendees` of the event.
             categories: The :attr:`categories` of the event.
             classification: The :attr:`classification` of the event.
             color: The :attr:`color` of the event.
@@ -429,6 +436,7 @@ class Event(Component):
             stamp: The :attr:`Component.stamp` of the event.
                 If None, this is set to the current time.
             start: The :attr:`start` of the event.
+            status: The :attr:`status` of the event.
             summary: The :attr:`summary` of the event.
             transparency: The :attr:`transparency` of the event.
             uid: The :attr:`uid` of the event.
@@ -464,6 +472,8 @@ class Event(Component):
         event.priority = priority
         event.transparency = transparency
         event.contacts = contacts
+        event.status = status
+        event.attendees = attendees
         if cls._validate_new:
             cls._validate_start_and_end(start, end)
         return event