+++ CHANGES.rst
@@ -35,6 +35,7 @@ New features:
 - Add ``Availability`` and ``Available`` components from :rfc:`7953`. See `Issue 654 <https://github.com/collective/icalendar/issues/654>`_.
 - Add ``stamp``, ``last_modified``, ``created``, ``CREATED``, ``busy_type``, ``class``, ``comments``, ``contacts``, ``location``, ``organizer``, ``priority``, and ``url`` properties to components that use them.
 - Add ``availabilities`` attribtue to ``Calendar``.
+- Add ``status``, ``transparency`` and ``attendees`` properties. See `Issue 841 <https://github.com/collective/icalendar/issues/841>`_.
 
 Bug fixes: